Extract filter defs from ObservationFilters to Filter module.
